-Removed "standard" npm dependency [that wasnt added intentionally] -fixed formatting errors +renamed files
