+++ api/api.js
@@ -19,8 +19,8 @@ function init (app) {
         var email = body.email
         var password = body.password
 
-        var validator = require('../src/inputvalidator')
-
+        var validator = require('../src/input_validator')
+        console.log(validator);
         response = {
           'valid-email': validator.isValidEmail(email),
           'valid-name': validator.isValidUserName(name),
