+++ html/input_validator_example.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <title>Example</title>
     <link type="text/css" rel="stylesheet" href="css/style.css"/>
-    <script type="text/javascript" src="js/inputvalidator.js"></script>
+    <script type="text/javascript" src="js/input_validator.js"></script>
     <script
         src="https://code.jquery.com/jquery-3.4.1.min.js"
         integrity="sha256-CSXorXvZcTkaix6Yvo6HppcZGetbYMGWSFlBw8HfCJo="
