+++ package.json
@@ -10,7 +10,6 @@
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "jquery": "^3.5.0",
-    "mysql": "^2.18.1",
-    "standard": "^14.3.3"
+    "mysql": "^2.18.1"
   }
 }
