+++ page/example.js
