+++ page/page.js
@@ -5,8 +5,8 @@ function init (app) {
     require('./example').page(req, res)
   })
 
-  app.get('/inputvalidator', function (req, res) {
-    require('./inputvalidator').page(req, res)
+  app.get('/input_validator', function (req, res) {
+    require('./input_validator').page(req, res)
   })
 
   app.get('/choose_login_method', function (req, res) {
